# Faster execution based on regex